requirements.txt
aiosmtplib==1.1.7
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
bcrypt==4.0.0
billiard==3.6.4.0
blinker==1.5
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
cssutils==2.6.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
emails==0.6
fastapi==0.85.0
fastapi-mail==1.2.0
flower==1.2.0
greenlet==1.1.3
httpx==0.23.0
humanize==4.4.0
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
premailer==3.10.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
pyasn1==0.4.8
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.4
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
starlette==0.20.4
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
